requirements.txt
matplotlib==3.8.4
numpy==2.1.2
opencv_python==4.10.0.84
opencv_python_headless==4.10.0.84
Pillow==11.0.0
streamlit==1.33.0
ultralytics==8.2.58
